refactor(work): own profile lifecycle in runtime
